\name{targetPredictorUpdater-package}
\alias{targetPredictorUpdater-package}
\alias{targetPredictorUpdater}
\docType{package}
\title{
What the package does (short line)
df sf sfd ssfdsfd
}
\description{
More about what it does (maybe more than one line)
blahblahblah
}
\details{
\tabular{ll}{
Package: \tab targetPredictorUpdater\cr
Type: \tab Package\cr
Version: \tab 0.99\cr
Date: \tab 2014-01-16\cr
License: \tab GPL 2 \cr
}
adsaasfd fa
}
\author{
Maciej Pajak

Maintainer: Maciej Pajak <m.pajak@sms.ed.ac.uk>

}
\references{
Literature or other references for background information
}
Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation
directory
\keyword{ package }
\seealso{
Optional links to other man pages, e.g.
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
simple examples of the most important functions
}